+++ spec/v1/providers/auth.spec.ts
@@ -23,9 +23,13 @@
 import { expect } from 'chai';
 import * as firebase from 'firebase-admin';
 
-import { CloudFunction, Event, EventContext } from '../../src/cloud-functions';
-import * as functions from '../../src/index';
-import * as auth from '../../src/providers/auth';
+import * as functions from '../../../src/index';
+import {
+  CloudFunction,
+  Event,
+  EventContext,
+} from '../../../src/v1/cloud-functions';
+import * as auth from '../../../src/v1/providers/auth';
 
 describe('Auth Functions', () => {
   const event: Event = {
