+++ tsconfig.release.json
@@ -10,5 +10,5 @@
     "target": "es2018",
     "typeRoots": ["./node_modules/@types"]
   },
-  "files": ["./src/index.ts", "./src/logger.ts", "./src/logger/compat.ts"]
+  "files": ["./src/index.ts", "./src/logger/index.ts", "./src/logger/compat.ts"]
 }
